git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@3655 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/stackwrap.c
@@ -4,8 +4,9 @@
 #include <kern/lib.h>
 #include <inc/setjmp.h>
 #include <inc/error.h>
+#include <inc/intmacro.h>
 
-#define STACKWRAP_MAGIC	0xabcd9262deed1713
+#define STACKWRAP_MAGIC	UINT64(0xabcd9262deed1713)
 
 struct stackwrap_state {
     stackwrap_fn fn;
